Refactor entity removal events to EntityRemoveListener
